Rename `render.tsx` to `promisemanager.tsx`
